Suggest some small tweaks to the repo (#3)
* Update .gitignore to exclude '.vscode/' folder

* Add '.sourcekit-lsp/config.json' to prioritize swift packet manager for contributors using Cursor/VSCode

* Rename apple.json to config.json for more semantic clarity

* Add sourcekit lsp config.json to the Example project

* Add 'Stream SourceKit Logs' to example app

Add this so we can easily stream sourcekit logs for the example app from within VSCode

* Add symlink_prefix build flag to Example app

Add this flag to show that the BSP still works, even with a custom symlink prefix flag.

* Update .gitignore to exclude '.vscode/' folder

* Fix aquery parsing

I noticed, the aquery parsing can run into an issue where a path contains both "external/" and "bazel-out/" — this commit refactors this logic to be more readable and use `continue` statements. We also check bazel-out/ first, as that would always precede external if both are used.